Refactor QuickToolbar to extract logic into smaller, maintainable hooks
Extracted session list management, session modal toggling, and route coach actions into separate custom hooks to reduce QuickToolbar's complexity.
